# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nutrition Assistant API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/weights": {
      "post": {
        "tags": [
          "weights"
        ],
        "summary": "Create Weight",
        "operationId": "create_weight_api_v1_weights_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeightLogCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Weight Api V1 Weights Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "weights"
        ],
        "summary": "List Weights",
        "operationId": "list_weights_api_v1_weights_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 30,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightLogListEnvelopeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/weights/{log_id}": {
      "patch": {
        "tags": [
          "weights"
        ],
        "summary": "Update Weight",
        "operationId": "update_weight_api_v1_weights__log_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Log Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeightLogUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Weight Api V1 Weights  Log Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "weights"
        ],
        "summary": "Delete Weight",
        "operationId": "delete_weight_api_v1_weights__log_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Log Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Weight Api V1 Weights  Log Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginationMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "total_pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Pages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "title": "PaginationMeta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeightLogCreateRequest": {
        "properties": {
          "logged_at": {
            "type": "string",
            "format": "date-time",
            "title": "Logged At"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Weight Kg"
          },
          "body_fat_percent": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Fat Percent"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "logged_at",
          "weight_kg"
        ],
        "title": "WeightLogCreateRequest"
      },
      "WeightLogListDataResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WeightLogListItemResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "WeightLogListDataResponse"
      },
      "WeightLogListEnvelopeResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "message_key": {
            "type": "string",
            "title": "Message Key"
          },
          "data": {
            "$ref": "#/components/schemas/WeightLogListDataResponse"
          },
          "error": {
            "type": "null",
            "title": "Error"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message_key",
          "data"
        ],
        "title": "WeightLogListEnvelopeResponse"
      },
      "WeightLogListItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "logged_at": {
            "type": "string",
            "title": "Logged At"
          },
          "weight_kg": {
            "type": "string",
            "title": "Weight Kg"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "logged_at",
          "weight_kg"
        ],
        "title": "WeightLogListItemResponse"
      },
      "WeightLogUpdateRequest": {
        "properties": {
          "weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "body_fat_percent": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Fat Percent"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "WeightLogUpdateRequest"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}